requirements.txt
snowflake-snowpark-python[pandas]
snowflake-cli-labs
python-dotenv